+++ src/main/java/com/greenfoxacademy/springwebapp/config/repositories/ProductRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductRepository extends CrudRepository<Product, Long> {
+public interface ProductRepository extends CrudRepository<Product, Integer> {
   List<Product> findAll();
-  Optional<Product> findById(Long id);
+  Optional<Product> findById(Integer id);
 
 }
